Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+1
> 
> Thanks Zack. I'll start testing and building on this. One question, how does the resulting tar file get added to pip? Is it managed with the VERSION file? or is it done manually?

I have Jenkins job that can push the package to PyPI once it's been merged, but currently that script triggers only when the VERSION file is a SemVer released version (ie no `-dev` or similar suffix).

I'm not sure if I've created a job for this specific repo, but I can create one. Did you want to test this on 0.x.x versions, then when closer to release create a 1.0.0 ?

Patch-set: 3