VOL-1460 Implemented a new approach to protos where go stubs are not committed to the repo. There are new instructions for clients consuming these protos in Go. Also removed vendor directory and dep, as it isn't really needed at this point.
Implemented a test to check if protos match up with go files. This is run with make test.
This works by generating them and checking this created any more untracked or unstaged changes.
It requires you to commit your changes locally to pass. This process to should be run by
jenkins to validate a commit. A consistent version of protoc must be used, which the Makefile checks for.

Change-Id: I8de28c515a9535e33909cbbd158fa046f96f3b1f