Update patch set 2
Patch Set 2:

For finding the google apis, there is another solution other than cloning the googleapi repo. We already include two of the them under protos/google/api, and those are all that are needed. The build script just has to be updated to point to this folder instead of the other one. I tested this and it worked, the files produced were the same according to git.

Patch-set: 2