Update patch set 2
Patch Set 2:

> (1 comment)

Yes, it has them. If we want to go get that instead, we could. But I think if we do that, we also want to address the fact that the python build needs them. I don't see the point if we keep them there for python protos.

Patch-set: 2
Reviewer: Gerrit User 1000455 <1000455@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>
Label: Code-Review=0