Update patch set 4
Patch Set 4:

> (1 comment)

My issue is that this is part of the make clean target. And my expectation from that target is that it removes build artifacts, and normally would apply to this directly. But this would mean running a make clean causes git status to report some files removed. So I would say having that done only as part of the test script makes sense.

Patch-set: 4
Label: Code-Review=0