Update patch set 4
Patch Set 4:

> I may have missed something, but it seems that some of the newly
 > added protos (inter_container.proto, afrouter.protos,
 > openolt.proto) have been deleted and we certainly need them with
 > the new core.  I suspect it may have just been an oversight.

If this change were to be merged it would be merged as a "v1" branch. Where the existing protos would be in a "v2" branch. That way from a Go app you could use gopkg.in to import either v1 or v2 protos.

It is unclear we want to move forward with this plan as go use gopkg.in that would have to be used as the go_packname in all the proto files. if we did that it does work.

my original experiment was to see if a single client could import and use both the v1 and v2 protos, the answer to that is no.

so now the question is what do we want to do. Do we want to support a go import style of "gopkt.in/opencord/voltha-protos.v1" and "gopkt.in/opencord/voltha-protos.v2"? or just leave things as they are?

And once we go "gopkt.in" it means it is used everywhere, i.e. "github.com/opencord/voltha-protos" won't work as the go import.

Patch-set: 4