Update patch set 4
Patch Set 4:

> Any reason why we are adding protos (bbf_fiber*)  to the repo?  
 > These files were used to support the XPON which has/will been
 > replaced by the tech profile in 2.x.

This was a bit of an experiment to see how/if we could get both protos (1.x and 2.x) into the same repo and use gopkg.in to allow Go apps to version between the two. I just grabbed all the protos from the 1.x tree for this test. We have to decide if we want to go forward with this or not. The versioning seems to work, although a single Go app importing and using both versions at the same time does not.

Patch-set: 4
Reviewer: Gerrit User 1000003 <1000003@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>
Label: Code-Review=0