Update patch set 9
Patch Set 9:

Hi Girish, this looks good.

Do you need to upload all the related files for this commit?

Since other adapters rely on the event proto then I suggest to mention it to the community (voltha discuss and TST) before we merge this change.  Any existing adapters, if any, that are using these protos will be impacted.

Thanks

Patch-set: 9