Update patch set 10
Patch Set 10:

> Hi Girish, this looks good.
 > 
 > Do you need to upload all the related files for this commit?
 > 
 > Since other adapters rely on the event proto then I suggest to
 > mention it to the community (voltha discuss and TST) before we
 > merge this change.  Any existing adapters, if any, that are using
 > these protos will be impacted.
 > 
 > Thanks

Ok I am going to drop a mail on voltha-discuss. 

I am not sure why the generated files are being maintained on the repo in the first place. When I compiled the protos, it looked like many generated files were not updated in a long time, so I am not sure if they should be updated. Do you have any clue about this. 

Another point on the VERSION of the voltha-protos. I had missed to bump of the version and I just changed from 0.1.4-dev to 0.2.0-dev. Is that Ok?

Patch-set: 10