Update patch set 2
Patch Set 2:

There is another review on voltha-protos https://gerrit.opencord.org/#/c/14766/ with updates to openolt.proto . Can that be part of this review too.
If we could have all the traffic profile and bandwidth profile related proto changes in a single review, it makes things easier to review, version and all the "logical" blocks could go in single commit.

Patch-set: 2