Update patch set 4
Patch Set 4:

(2 comments)

I think the question is do we want to solve the logging issue for the system or patch so it works for the cores. Currently, this draft addresses the "core" issue not log settings for the "system".

Patch-set: 4
Reviewer: Gerrit User 1000003 <1000003@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>
Label: Code-Review=0