Update patch set 3
Patch Set 3:

> openolt agent uses the local protos which is there on the same
 > repository.
 > The voltha 2.x components uses protos from the voltha-protos
 > repository which is causing synchronization issues .
 > Do we need ensure that we should have a common repo for PROTOS
 > which will be used by all the components?

I agree that we have to keep the agent's local protos and the adapter's protos in sync.  And i also agree that the agent should no longer have its own copy of protos as it causes sync issues like this.

My only point is that if this change is going to be cherry-picked into adapter 2.1, then the registration_id proto addition would go with it.  The goal of 2.1 is to be compatible with agent 1.7 and agent 1.7 branch does not have registration_id in its proto.  If this commit were broken into two commits, the registration_id change could stay in master while the disable/enable change could be in master and cherry-picked into 2.1

Patch-set: 3