Update patch set 3
Patch Set 3:

> > I think more discussion is needed in the community about event
 > > filtering. As far as I under stood, the voltha system
 > > (adapters/core/kafka etc) is just supposed to act as an event
 > > conduit and not really filter anything. So, I am not sure where
 > > this is coming from, unless I am missing something.
 > 
 > This item depends on the user-specific requirements if someone
 > wants to suppress some events. If no filters are created there
 > won't be any filtration.

This may be a good feature to have. But since this a major functionality at the VOLTHA, it might be worth to solicit feedback from the community. From earlier discussions it was decided that VOLTHA acts as a event conduit, but now we want to do filtering too. So, it better to solicit feedback IMHO.

Patch-set: 3
Reviewer: Gerrit User 1000127 <1000127@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>
Label: Code-Review=0