Update patch set 3
Patch Set 3:

> (1 comment)

I also tempted to put it into TrafficSchedulers. However, by definition of this task, a tech_profile_id should differentiate a scheduler. If we put tech_profile_id in the parent structure and if we want to use CreateTrafficSchedulers to create a group of schedulers representing separate tech profiles, we will not be able to do that. That's why, I put tech_profile_id in each of the TrafficScheduler structure.

Patch-set: 3
Reviewer: Gerrit User 1000558 <1000558@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>
Label: Code-Review=0