Update patch set 3
Patch Set 3:

> Using a UUID for the port id implies that on a delete/add of a
 > device the same port will be represented by a different UUID each
 > time.   Is that the intent?  Not sure how this benefits the NB
 > system.

We were thinking in the same lines as the UUID for a device. Understanding is that once a device is removed from voltha all context about that device is lost and re-add is as good as a new device, is this understanding correct? Can't the ports be treated the same way?

Patch-set: 3