Update patch set 3
Patch Set 3:

> > Using a UUID for the port id implies that on a delete/add of a
 > > device the same port will be represented by a different UUID each
 > > time.   Is that the intent?  Not sure how this benefits the NB
 > > system.
 > 
 > We were thinking in the same lines as the UUID for a device.
 > Understanding is that once a device is removed from voltha all
 > context about that device is lost and re-add is as good as a new
 > device, is this understanding correct? Can't the ports be treated
 > the same way?

Yes, Amit, they definitely can be treated the same way.  My comment is more about how will this benefit a NB system to have that added ID.    I am thinking more around what makes a device unique (and its ports) across restarts.  This mostly applies to ONU devices.  In the Core whenever an ONU is being added we go over a number of parameters (e.g. onu-id, serial number, vendor id, etc) to see if we are adding a duplicate.  I don't know if it would be beneficial to have a unique hash for a given device that does not change across restarts/delete.  In the Core, we could use the device hash as a key to internal data structures hence when adding an ONU device it's just one map lookup to see if that device exists.   

It's a good idea to discuss more at the TST.

Patch-set: 3
Reviewer: Gerrit User 1000056 <1000056@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>
Label: Code-Review=0