Update patch set 10
Patch Set 10:

If I am reading the diagrams on the referenced JIRA correctly, both options (FORCED and non-FORCED) send a request to the adapter and then wait for a response. When the message is sent to the adapter the transient state for the device in the core in both use cases seems to be "DELETING_FROM_ADAPTER" when starting from a non-"PREPROVISIONED" state. 

When a response is received from the adapter different actions are taken depending if this is a forced delete or not, but I am not seeing how the core can tell the difference at this point as the transient state is "DELETING_FROM_ADATPER" in both cases.

Starting to wonder if the transient state needs to be a bitset, so that in the forced case the value can be "FORCED_DELETED | DELETING_FROM_ADAPTER" where in the non-forced case it is just "DELETING_FROM_ADAPTER". Am I missing something?

Patch-set: 10
Reviewer: Gerrit User 1000003 <1000003@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>