Update patch set 10
Patch Set 10:

In the forced delete case the core is waiting for a response from the adapter before actually deleting the device. One of the use cases where force delete was expected to be needed was when the adapter or physical device may be non-reachable.

If the adapter gets a delete device for a device it does not recognize what is its behavior? Does it respond with an error to the core? Does it simply drop the message?

My concern is that in some situations the core may be waiting for a time out as opposed to just directly processing the forced delete immediately. Can this be worked to that a message is sent to the adapter, but the core proceeds to forced delete immediately anyway and then simply ignores or logs the message for a device that comes back from an adapter?

Patch-set: 10
Reviewer: Gerrit User 1000003 <1000003@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>