Update patch set 10
Patch Set 10:

> Patch Set 10: Code-Review+1
> 
> I'm fine with the changes, I would like to hold off on the move to this version of the protos after 2.6. Thoughts ?

These changes are relevant to the transaction error handling which has been worked out for release 2.6.  Relevant changes are in progress in the rw-core and ofagent using this set of protos.  I was waiting for the rw-core and ofagent code to firm up on their proto need to submit these changes.

Patch-set: 10