Update patch set 4
Patch Set 4:

(33 comments)

Sorry for so many comments, tried to express my problems of understanding what is really intended here. I think it would be easier if the handling for downloadToAdapter and downloadToDevice (including their indications and states) would further be separated (as already kind of started here).
Prrobably no need to answer all comments, maybe the planned discussion on this makes it clearer.

Patch-set: 4