Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> @Girish Should we take port no as input request and get onu id and intf id on based of that? I guess that would be better?

No, we should not take port no. Port no could change based on uni_id of the ONU also. But here we are collecting counters of the ONU PON which is collection of to-and-fro traffic from all the uni ports of the ONU. So, onu id and intf id should suffice and is right.

Patch-set: 1
Reviewer: Gerrit User 1000127 <1000127@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>