Update patch set 1
Patch Set 1: Code-Review+1

There is an existing FAILED state which means the device is unrecoverable and require an operator action.   RECONCILE_FAILED, while stating which action placed the device in a failed state, will also result in an operator action.  Not sure if there are other conditions that would put the device in a failed state (the current FAILED state is actually not used).  Food for thoughts for the future.

Patch-set: 1
Reviewer: Gerrit User 1000056 <1000056@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>
Label: Code-Review=+1